# Conflicts:
#	.gitignore
#	bin/log1.txt
#	bin/mbs.exe
#	bin/mbs_d.exe
#	pro/BeamSplitting.pro
#	pro/BeamSplitting.pro.user
#	pro/MBS.pri
#	src/Beam.h
#	src/Tracer.cpp
#	src/Tracer.h
#	src/TracerBackScatterPoint.cpp
#	src/TracerGO.cpp
#	src/common/Beam.cpp
#	src/geometry/Polygon.cpp
#	src/geometry/Polygon.h
#	src/geometry/geometry_lib.cpp
#	src/geometry/geometry_lib.h
#	src/handler/Handler.cpp
#	src/handler/Handler.h
#	src/handler/HandlerGO.cpp
#	src/handler/HandlerGO.h
#	src/handler/HandlerPO.cpp
#	src/handler/HandlerPO.h
#	src/handler/HandlerTotalGO.cpp
#	src/handler/HandlerTotalGO.h
#	src/handler/HandlerTracksGO.cpp
#	src/handler/HandlerTracksGO.h
#	src/main.cpp
#	src/particle/Bullet.cpp
#	src/particle/CertainAggregate.cpp
#	src/particle/HexagonalAggregate.cpp
#	src/particle/Particle.cpp
#	src/particle/Particle.h
#	src/scattering/Scattering.cpp
#	src/scattering/Scattering.h
#	src/scattering/ScatteringConvex.cpp
#	src/scattering/ScatteringNonConvex.cpp
#	tool/StlToCry/Converter.cpp
#	tool/StlToCry/Converter.h
#	tool/StlToCry/StlToCry.pro
#	tool/StlToCry/main.cpp
#	tool/VoronovParticleGenerator/VoronoiLattice.cpp
#	tool/VoronovParticleGenerator/VoronoiLattice.h
#	tool/VoronovParticleGenerator/main.cpp

+++ src/handler/HandlerBackscatterPoint.h
@@ -5,15 +5,16 @@
 class HandlerBackScatterPoint : public HandlerPO
 {
 public:
-    HandlerBackScatterPoint(Particle *particle, Light *incidentLight, float wavelength = 0);
+	HandlerBackScatterPoint(Particle *particle, Light *incidentLight,
+							double wavelength = 0);
 
-    void HandleBeams(std::vector<Beam> &beams) override;
-    void SetTracks(Tracks *tracks) override;
+	void HandleBeams(std::vector<Beam> &beams) override;
+	void SetTracks(Tracks *tracks) override;
 
-    void OutputContribution(ScatteringFiles &files, double angle, double energy,
-                            bool isOutputGroups, std::string prefix = "");
-
+	void OutputContribution(ScatteringFiles &files, double angle, double energy,
+							bool isOutputGroups, std::string prefix = "");
 private:
-    PointContribution *originContrib;
-    PointContribution *correctedContrib;
+	PointContribution *originContrib;
+	PointContribution *correctedContrib;
 };
+
